salebot/pyproject.toml
[project]
name = "salebot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.86.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "python-telegram-bot>=22.7",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-mock>=3.14.0",
    "respx>=0.22.0",
    "ruff>=0.15.7",
]
